new package for e2e testing
